chore: expose LedgerOpts for tests
